refactor: line dao에 데이터 전달 시 dto 이용하도록 변경
